+++ src-tauri/framework/wry/src/webkitgtk/synthetic_mouse_events.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use std::{cell::RefCell, rc::Rc};
 
 use gtk::{